Move certificate decode method out of the eval class
