+++ sdk/src/androidTest/java/ch/admin/bag/covidcertificate/eval/CertificateDecoderTest.kt
@@ -11,16 +11,17 @@
 package ch.admin.bag.covidcertificate.eval
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
+import ch.admin.bag.covidcertificate.eval.decoder.CertificateDecoder
 import org.junit.Assert.assertTrue
 import org.junit.Test
 import org.junit.runner.RunWith
 
 @RunWith(AndroidJUnit4::class)
-class EvalTest {
+class CertificateDecoderTest {
 
 	@Test
 	fun testDecode() {
-		val decodeState = Eval.decode(HC1_A)
+		val decodeState = CertificateDecoder.decode(HC1_A)
 		assertTrue(decodeState is DecodeState.SUCCESS)
 	}
 }
